Added search for deleted users.

+++ src/Web/EssayCompetition.Web/Areas/Administration/Controllers/RolesController.cs
@@ -15,7 +15,7 @@
 
     public class RolesController : AdministrationController
     {
-        private const int PageSize = 5;
+        private const int PageSize = 10;
         private readonly IUsersService usersService;
         private readonly IRolesService rolesService;
         private readonly UserManager<ApplicationUser> userManager;
@@ -36,8 +36,8 @@
             viewModel.Pager ??= new PagerViewModel();
             viewModel.Pager.CurrentPage = viewModel.Pager.CurrentPage <= 0 ? 1 : viewModel.Pager.CurrentPage;
 
-
-            viewModel.Users = this.usersService.GetUsersWithRoles<UserViewModel>(viewModel.Pager.CurrentPage, PageSize, viewModel.SearchString, viewModel.SortOrder);
+            viewModel.Users = this.usersService.GetUsersWithRoles<UserViewModel>
+                (viewModel.Pager.CurrentPage, PageSize, viewModel.SearchString, viewModel.SortOrder, viewModel.SearchOnlyDeleted);
 
             foreach (var user in viewModel.Users)
             {
@@ -88,15 +88,15 @@
             return this.View(viewModel);
         }
 
-        public IActionResult Details(string id)
+        public IActionResult Details(string id, bool fromDeleted)
         {
-            if (!this.usersService.HasUserWithId(id))
+            if (!this.usersService.HasUserWithId(id) && !this.usersService.HasDeletedUserWithId(id))
             {
                 return this.NotFound();
             }
 
             var viewModel = this.usersService.GetUserById<DetailsViewModel>(id);
-
+            viewModel.FromDeleted = fromDeleted;
             return this.View(viewModel);
         }
 
@@ -152,7 +152,23 @@
 
         public async Task<IActionResult> Delete(string id)
         {
+            if (!this.usersService.HasUserWithId(id))
+            {
+                return this.NotFound();
+            }
+
             await this.usersService.DeleteUserAsync(id);
+            return this.RedirectToAction("Index");
+        }
+
+        public async Task<IActionResult> UnDelete(string id)
+        {
+            if (!this.usersService.HasDeletedUserWithId(id))
+            {
+                return this.NotFound();
+            }
+
+            await this.usersService.UnDeleteUserAsync(id);
             return this.RedirectToAction("Index");
         }
     }
